// src/types/CurrencyType.ts
// type CurrencyType = {
// 	[key: string]: {
// 		CharCode: string
// 		ID: string
// 		Name: string
// 		Nominal: number
// 		NumCode: string
// 		Previous: number
// 		Value: number
// 	}
// }

type CurrencyType = {
	CharCode: string
	ID: string
	Name: string
	Nominal: number
	NumCode: string
	Previous: number
	Value: number
}

export default CurrencyType